# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Local API Testing Laboratory",
    "description": "Enterprise-style local API testing laboratory.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users (Admin/Manager)",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Users retrieved" } }
      }
    },
...
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Get all products",
        "security": [],
        "parameters": [
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Products retrieved" } }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Create a new product (Admin/Manager)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "price": { "type": "number" },
                  "stock": { "type": "integer" },
                  "category": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Product created" } }
      }
    },
...
    "/api/orders": {
      "post": {
        "tags": ["Orders"],
        "summary": "Create an order",
        "parameters": [
          { "name": "Idempotency-Key", "in": "header", "required": false, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "items": { "type": "array", "items": { "type": "object" } },
                  "totalAmount": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Order created" } }
      },
      "get": {
        "tags": ["Orders"],
        "summary": "Get user orders",
        "responses": { "200": { "description": "Orders retrieved" } }
      }
    },
...
    "/api/upload": {
      "post": {
        "tags": ["Uploads"],
        "summary": "Upload a single file",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "File uploaded" } }
      }
    },
...
    "/api/jobs": {
      "post": {
        "tags": ["Jobs"],
        "summary": "Create an async background job",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "type": { "type": "string" } } } } }
        },
        "responses": { "202": { "description": "Job queued" } }
      }
    },
    "/api/jobs/{id}/status": {
      "get": {
        "tags": ["Jobs"],
        "summary": "Get job status",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Status retrieved" } }
      }
    },
...
    "/api/feature-flags": {
      "get": {
        "tags": ["Feature Flags"],
        "summary": "Get feature flags",
        "responses": { "200": { "description": "Flags retrieved" } }
      },
      "patch": {
        "tags": ["Feature Flags"],
        "summary": "Update feature flags (Admin)",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "flag": { "type": "string" }, "enabled": { "type": "boolean" } } } } }
        },
        "responses": { "200": { "description": "Flag updated" } }
      }
    },
...
    "/api/test/slow": {
      "get": {
        "tags": ["Testing / Debug"],
        "summary": "Test slow response",
        "security": [],
        "parameters": [
          { "name": "delay", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Delayed response" } }
      }
    },
...
    "/api/test/cpu-load": {
      "get": {
        "tags": ["Testing / Debug"],
        "summary": "Test CPU load",
        "security": [],
        "parameters": [
          { "name": "ms", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "CPU task completed" } }
      }
    },
...
    "/api/test/heavy-response": {
      "get": {
        "tags": ["Testing / Debug"],
        "summary": "Test huge payload generation",
        "security": [],
        "parameters": [
          { "name": "size", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Huge JSON returned" } }
      }
    }
  }
}